Change GetSkillDmgAmt from Int16 to Int32 (SkillDamageAmount2 which uses function is Int32)

+++ zone/mob.cpp
@@ -5473,7 +5473,7 @@ int16 Mob::GetSkillReuseTime(uint16 skill)
 	return skill_reduction;
 }
 
-int16 Mob::GetSkillDmgAmt(uint16 skill)
+int32 Mob::GetSkillDmgAmt(uint16 skill)
 {
 	int skill_dmg = 0;
